+++ src/main/kotlin/me/ffl/intellijDirectoryTests/Inline.kt
@@ -1,5 +1,6 @@
 package me.ffl.intellijDirectoryTests
 
+import com.intellij.openapi.application.runWriteAction
 import me.ffl.intellijDirectoryTests.MarkupFile.Companion.findCarets
 import kotlin.io.path.div
 
@@ -9,6 +10,8 @@ val inlineExecutor: KotestExecutor = {
     val (caretFile, caretOffset) = checkNotNull(beforeFiles.findCarets().singleOrNull()) {
         "Inline tests must only have one caret"
     }
-    caretFile.inlineAt(caretOffset)
+    runWriteAction {
+        caretFile.inlineAt(caretOffset)
+    }
     checkAfterProject()
 }