+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+- File content difference was not shown in the error message, when comparing files to the expected file after a mutating operation
+
 ## [0.2.1] 2023-07-30
 
 ### Fixed